requirements.txt
# Web Framework
Flask==1.1.2

# OpenAI API
openai==0.27.0

# Image Processing
opencv-python==4.5.3.56
Pillow==8.3.2

# Database
psycopg2==2.9.1

# Server
gunicorn==20.1.0

# Other
requests==2.26.0
numpy==1.21.2
<FINAL_CODE>